Refactor VerificationQRCodeProcessor
